requirements.txt
pytube3==9.6.4
tqdm==4.45.0
beautifulsoup4==4.9.0
